+++ R/dwh.R
@@ -1,6 +1,3 @@
-#============================================================================
-#  function: dwh_connect
-#============================================================================
 #' connect to DWH
 #'
 #' connect to datawarehouse (DWH) using ODBC
@@ -19,24 +16,12 @@
 
 dwh_connect <- function(dsn, user = NA, pwd = NA, pwd_crypt = FALSE, ...)  {
 
-  if (is.na(user))  {
-    # use single sign on
-    channel <- DBI::dbConnect(odbc::odbc(), dsn, ...)
+  .Defunct(msg = paste(
+    "All dwh_ functions are no longer included in {explore}.",
+    "You find the source code at https://github.com/rolkra/dwh"))
 
-  } else {
-    # use user & passwort
-    channel <- DBI::dbConnect(odbc::odbc(), dsn,
-                              user = user,
-                              password = if (pwd_crypt == TRUE) decrypt(pwd) else pwd,
-                              ...
-    )
-  } # if
-  return(channel)
 }
 
-#============================================================================
-#  function: dwh_disconnect
-#============================================================================
 #' disconnect from DWH
 #'
 #' disconnect from datawarehouse (DWH) using a ODBC connection
@@ -50,12 +35,13 @@ dwh_connect <- function(dsn, user = NA, pwd = NA, pwd_crypt = FALSE, ...)  {
 #' @export
 
 dwh_disconnect <- function(connection, ...)  {
-  DBI::dbDisconnect(connection, ...)
+
+  .Defunct(msg = paste(
+    "All dwh_ functions are no longer included in {explore}.",
+    "You find the source code at https://github.com/rolkra/dwh"))
+
 }
 
-#============================================================================
-#  function: dwh_read_table
-#============================================================================
 #' read a table from DWH
 #'
 #' read a table from DWH using a ODBC connection
@@ -73,21 +59,12 @@ dwh_disconnect <- function(connection, ...)  {
 
 dwh_read_table <- function(connection, table, names_lower = TRUE, ...)  {
 
-  # define sql
-  sql <- paste0("select * from ", table)
+  .Defunct(msg = paste(
+    "All dwh_ functions are no longer included in {explore}.",
+    "You find the source code at https://github.com/rolkra/dwh"))
 
-  # read data from dwh
-  data <- DBI::dbGetQuery(connection, sql, ...)
-
-  # convert names to lower case
-  if (names_lower) names(data) <- tolower(names(data))
-
-  return(data)
 }
 
-#============================================================================
-#  function: dwh_read_data
-#============================================================================
 #' read data from DWH
 #'
 #' read data from DWH using a ODBC connection
@@ -105,18 +82,12 @@ dwh_read_table <- function(connection, table, names_lower = TRUE, ...)  {
 
 dwh_read_data <- function(connection, sql, names_lower = TRUE, ...)  {
 
-  # read data from dwh
-  data <- DBI::dbGetQuery(connection, sql, ...)
+  .Defunct(msg = paste(
+    "All dwh_ functions are no longer included in {explore}.",
+    "You find the source code at https://github.com/rolkra/dwh"))
 
-  # convert names to lower case
-  if (names_lower) names(data) <- tolower(names(data))
-
-  return(data)
 }
 
-#============================================================================
-#  function: dwh_fastload
-#============================================================================
 #' write data to a DWH table
 #'
 #' write data fast to a DWH table using a ODBC connection
@@ -138,25 +109,8 @@ dwh_read_data <- function(connection, sql, names_lower = TRUE, ...)  {
 
 dwh_fastload <- function(data, dsn, table, overwrite = FALSE, append = FALSE, ...)  {
 
-  # check table (must be 'database.table')
-  # split string at '.'
-  table_split <- strsplit(table, split="[.]")
-  database_name <- table_split[[1]][1]
-  table_name <- table_split[[1]][2]
-
-  # valid database_name and table_name?
-  if ( is.na(database_name) | is.na(table_name) )   {
-    stop("table must be in the format 'database.table'")
-  }
-  stopifnot (nchar(database_name) > 0, nchar(table_name) > 0)
-
-  # connect
-  con <- DBI::dbConnect(odbc::odbc(), dsn=dsn, database=database_name, ...)
-
-  # write data
-  DBI::dbWriteTable(con, name=table_name, value=data, overwrite=overwrite, append=append)
-
-  # disconnect
-  DBI::dbDisconnect(con)
+  .Defunct(msg = paste(
+    "All dwh_ functions are no longer included in {explore}.",
+    "You find the source code at https://github.com/rolkra/dwh"))
 
 } # dwh_fastload